update product and prescription logic and add image upload
- Add UploadController for Cloudinary image management.
- Enhance Product and Prescription entities with more precise data types (BigDecimal) and additional fields like sub-category and temple length.
- Refactor product logic to handle variant-specific refractive indices and automatic base price calculation.
- Update data.sql to reflect schema changes and expand seed data for lenses and accessories.

+++ src/main/java/com/swp391/cclearly/dto/prescription/SavePrescriptionRequest.java
@@ -1,5 +1,6 @@
 package com.swp391.cclearly.dto.prescription;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,19 +8,19 @@ public class SavePrescriptionRequest {
   private String orderItemId; // UUID of the order item
 
   // Right eye (OD)
-  private Float sphOd;
-  private Float cylOd;
-  private Float axisOd;
-  private Float addOd;
+  private BigDecimal sphOd;
+  private BigDecimal cylOd;
+  private Integer axisOd;
+  private BigDecimal addOd;
 
   // Left eye (OS)
-  private Float sphOs;
-  private Float cylOs;
-  private Float axisOs;
-  private Float addOs;
+  private BigDecimal sphOs;
+  private BigDecimal cylOs;
+  private Integer axisOs;
+  private BigDecimal addOs;
 
   // PD
-  private Float pd;
+  private BigDecimal pd;
 
   // Prescription image/notes
   private String imageUrl;
